feat(venue): rename ContactBlockNew > ContactBlock
