+++ src/features/venue/components/PracticalInformation.tsx
@@ -3,7 +3,7 @@ import { FlatList, ListRenderItemInfo } from 'react-native'
 import styled from 'styled-components/native'
 
 import { VenueResponse } from 'api/gen'
-import { ContactBlock } from 'features/venue/components/ContactBlockNew/ContactBlockNew'
+import { ContactBlock } from 'features/venue/components/ContactBlock/ContactBlock'
 import { NoInformationPlaceholder } from 'features/venue/components/NoInformationPlaceholder'
 import { AccessibilityBlock } from 'ui/components/accessibility/AccessibilityBlock'
 import { Separator } from 'ui/components/Separator'
